requirements.txt
autogen-agentchat>=0.4.0
autogen-ext>=0.4.0
autogen-core>=0.4.0
streamlit>=1.30.0
python-dotenv>=1.0.0
openai>=1.0.0
